added contributor feed file and navigation bar file, changes made to Search CSS, search, dashboard
