+++ src/App.js
@@ -7,6 +7,7 @@ import Account from './components/scripts/account';
 import Search from './components/scripts/search';
 import FormComponent from './FormContributor.js';
 import Insights from "./components/scripts/insights";
+import ContributorPage from "./components/scripts/ContributorFeed";
 import { OpenAI } from './openAI'
 
 function App() {
@@ -22,6 +23,7 @@ function App() {
         <Route path="/contributor" element = {<FormComponent />} />
         <Route path="/account" element = {<Account />} />
         <Route path="/insights" element = {<Insights />} />
+        <Route path="/contributorfeed" element={<ContributorPage/>}/>
       </Routes>
     </Router>
   );
