+++ src/components/scripts/dashboard.js
@@ -30,42 +30,7 @@ import shadows from "@mui/material/styles/shadows";
 import styles from "../styles/Dashboard.module.css";
 import Insights from './insights'
 import quote1 from '../ArticleImage/resources/quote1.jpg'
-
-const Search = styled('div')(({ theme }) => ({
-  position: 'relative',
-  borderRadius: theme.shape.borderRadius,
-  backgroundColor: alpha(theme.palette.common.white, 0.15),
-  '&:hover': {
-    backgroundColor: alpha(theme.palette.common.white, 0.25),
-  },
-  marginLeft: 0,
-  width: '100%',
-  [theme.breakpoints.up('sm')]: {
-    marginLeft: theme.spacing(1),
-    width: 'auto',
-  },
-}));
-
-const SearchIconWrapper = styled('div')(({ theme }) => ({
-  padding: theme.spacing(0, 2),
-  height: '100%',
-  position: 'absolute',
-  pointerEvents: 'none',
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'center',
-  color: '#3A448C'
-
-}));
-
-const StyledPaper = styled(Paper)(({ theme }) => ({
-  display: 'flex',
-  alignItems: 'center',
-  width: '100%',
-  borderRadius:'50px',
-  margin: theme.spacing(1),
-
-}));
+import NaviBar  from "./navigationBar";
 
 const Styled2Paper = styled(Paper)(({ theme }) => ({
   display: 'relative',
@@ -77,30 +42,6 @@ const Styled2Paper = styled(Paper)(({ theme }) => ({
   backgroundColor:'white'
 }));
 
-
-
-const StyledInputBase = styled(InputBase)(({ theme }) => ({
-  color: 'inherit',
-  '& .MuiInputBase-input': {
-    padding: theme.spacing(1, 1, 1, 0),
-    // vertical padding + font size from searchIcon
-    width: '100%',
-    paddingLeft: `calc(1em + ${theme.spacing(4)})`,
-    transition: theme.transitions.create('width'),
-    [theme.breakpoints.up('sm')]: {
-      width: '20ch',
-      '&:focus': {
-        width: '30ch',
-      },
-    },
-  },
-}));
-
-const StyledToolbar = styled(Toolbar)(({ theme }) => ({
-  display: 'flex',
-  alignItems: 'center',
-  padding: theme.spacing(1, 2),
-}));
 
 const CustomTypography = styled(Typography)(({ theme }) => ({
   fontFamily: 'Montserrat, sans-serif',
@@ -123,17 +64,7 @@ function Dashboard() {
     const [fullContent, setFullContent] = useState([])
     // used as to pass query to search page
     const [searchInput, setSearchInput] = useState("");
-    const [searchFocused, setSearchFocused] = useState(false);
 
-    const handleSearchFocus = () => {
-      setSearchFocused(true);
-    };
-  
-    const handleSearchBlur = () => {
-      setSearchFocused(false);
-    };
-  
-    
     
     const [state, setState] = React.useState({
       top: false,
@@ -150,36 +81,7 @@ function Dashboard() {
       setState({ ...state, [anchor]: open });
     };
   
-    const list = (anchor) => (
-      <Box
-        sx={{ width: anchor === 'top' || anchor === 'bottom' ? 'auto' : 250 }}
-        role="presentation"
-        onClick={toggleDrawer(anchor, false)}
-        onKeyDown={toggleDrawer(anchor, false)}
-      >
-        <List>
-          {['Account', 'Contributor'].map((text, index) => (
-            <ListItem key={text} disablePadding>
-              <ListItemButton onClick={()=>navigate(`/${text.toLowerCase()}`)}>
-                <ListItemIcon>
-                  {index % 2 === 0 ? <InboxIcon /> : <MailIcon />}
-                </ListItemIcon>
-                <ListItemText primary={text} />
-              </ListItemButton  >
-            </ListItem>
-          ))}
-        </List>
-        <Divider />
-      </Box>
-    );
   
-    // on submit, we navigate to the search page and feed it the searchInpu\t
-    const handleSubmit = (event) =>{
-      event.preventDefault()
-      console.log(searchInput)
-      navigate('/search', { state:
-        {query: searchInput}});
-    }
      // on page load, we grab the current user's information and populate our variables
     useEffect(() => {
         const unsubscribe = auth.onAuthStateChanged(async (user) => {
@@ -203,43 +105,6 @@ function Dashboard() {
         return unsubscribe;
       }, [auth, firestore]);
 
-    // When topics array has been filled from the previous useEffect, then we will begin querying articles from the database 
-    // based off of their links
-    // We create an object, which holds an article's content, id, title, descritpion, and likes
-    /*
-    useEffect(() => {
-      const fetchArticles = async () => {
-        const newContent = [];
-        for (const topic of topics) {
-          console.log(topic)
-          try {
-            const q = query(collection(db, "articles"), where("topic", "==", topic), limit(1));
-            const querySnapshot = await getDocs(q);
-            querySnapshot.forEach((doc) => {
-              // doc.data() is never undefined for query doc snapshots
-              console.log(doc.data())
-              const article = {
-                id: doc.id,
-                title: doc.data().title,
-                description: doc.data().url,
-                content: doc.data().content,
-                userLikes: [],
-                likes: doc.data().userLikes.length,
-              }
-              newContent.push(article);;
-            });
-          } catch (error) {
-            console.error(`Error fetching articles for ${topic}`, error);
-          }
-        }
-        setFullContent(newContent);
-      };
-  
-      if (topics.length) {
-        fetchArticles();
-      }
-    }, [topics]);
-    */
     useEffect(() => {
       const fetchArticles = async () => {
         const newContent = [];
@@ -275,62 +140,9 @@ function Dashboard() {
     }, [topics]);
 
 
-    // When the input in the search field is changed, the update will be reflected in our variable
-    const handleChange = (e) => {
-      e.preventDefault();
-      setSearchInput(e.target.value);
-    };
-
-
     return (
       <main > 
-         <div>
-      {['left', 'right', 'top', 'bottom'].map((anchor) => (
-        <React.Fragment key={anchor}>
-          <Drawer
-            anchor={anchor}
-            open={state[anchor]}
-            onClose={toggleDrawer(anchor, false)}
-          >
-            {list(anchor)}
-          </Drawer>
-        </React.Fragment>
-      ))}
-    </div>
-          <Box sx={{ flexGrow: 1 }}>
-      <AppBar position="static" sx={{ background: 'linear-gradient(to right, #A473E6, #9DB7FD, #87FAE7)', shadow:'none', height:'80px'}}>
-        <StyledToolbar>
-          <IconButton
-            size="large"
-            edge="start"
-            color="inherit"
-            aria-label="open drawer"
-            sx={{ mr: 2 }}
-            onClick={toggleDrawer('left', true)}
-          >
-            <MenuIcon />
-          </IconButton>
-          <Search>
-
-          <StyledPaper component="form" onSubmit={handleSubmit}>
-          <SearchIconWrapper>
-            <SearchIcon />
-          </SearchIconWrapper>
-            <StyledInputBase
-              placeholder="Search…"
-              inputProps={{ 'aria-label': 'search' }}
-              value={searchInput}
-              onChange={(event) => setSearchInput(event.target.value)}
-            />
-          </StyledPaper>
-          </Search>
-          <div className= {styles.logos}> 
-            <h1 className={styles.logo1}>Stake The Stigma</h1>
-            <h2 className={styles.logo2}>Destigmatizing Women's Health.</h2>
-          </div>
-        </StyledToolbar>
-      </AppBar>
-    </Box>
+         <NaviBar/>
         <div >
         {/* <h1> Stake The Stigma.</h1>
         <h2 > Destigmatizing Women's Health</h2> */}
